refactor(card): colocate deck-editor orchestration with its view
Move the deck-editor-only composables (list-controller + cardEditorKey,
actions, bulk-actions, virtual-list) and the deck view-shell into
views/deck/composables/, colocated with the only feature that consumes
them. composables/card/ now holds reusable, feature-neutral primitives
only (mutations, selection, prompts, gates, image upload/dropzone).

Establishes the convention: feature-specific composables live with their
feature; the shared composables/ tree is for cross-feature primitives.
