+++ src/utils/card-editor/selection-payload.ts
@@ -1,4 +1,5 @@
-import type { CardSelection, VirtualCardList } from '@/composables/card'
+import type { CardSelection } from '@/composables/card'
+import type { VirtualCardList } from '@/views/deck/composables'
 
 export type DeleteArgs = { except_ids: number[] } | { cards: Card[] }
