+++ src/views/deck/card-editor/card-face-uploader.vue
@@ -6,7 +6,8 @@ import UiIcon from '@/components/ui-kit/icon.vue'
 import UiTooltip from '@/components/ui-kit/tooltip.vue'
 import FaceImageDropzone from './face-image-dropzone.vue'
 import UiButton from '@/components/ui-kit/button.vue'
-import { cardEditorKey, CARD_IMAGE_MAX_BYTES, useFaceImageUpload } from '@/composables/card'
+import { CARD_IMAGE_MAX_BYTES, useFaceImageUpload } from '@/composables/card'
+import { cardEditorKey } from '@/views/deck/composables'
 import { cardImageUrl } from '@/api/media'
 import { CARD_ATTRIBUTES_DEFAULTS } from '@/utils/deck/defaults'
 import { emitSfx } from '@/sfx/bus'
