+++ src/views/deck/card-grid/empty-state.vue
@@ -5,8 +5,8 @@ import UiIcon from '@/components/ui-kit/icon.vue'
 import { computed, inject } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useMatchMedia } from '@/composables/ui/media-query'
-import { type CardGridSize } from '@/composables/deck/view-shell'
-import { cardEditorKey } from '@/composables/card'
+import { type CardGridSize } from '@/views/deck/composables/view-shell'
+import { cardEditorKey } from '@/views/deck/composables'
 
 const { t } = useI18n()
