+++ src/views/deck/composables/index.ts
@@ -1,0 +1,14 @@
+// Deck-view card-editor orchestration. Colocated with the deck view because
+// it's consumed only by views/deck/** (card-editor, card-grid, mode-toolbar,
+// deck-hero). For reusable, feature-neutral card primitives (mutations,
+// selection, prompts, gates, image upload) import from `@/composables/card`.
+
+export { useCardListController, cardEditorKey, type CardListController } from './list-controller'
+export { useCardActions, type CardActions } from './actions'
+export { useBulkActions } from './bulk-actions'
+export {
+  useVirtualCardList,
+  type VirtualCardList,
+  type CardWithClientId,
+  type CardEntry
+} from './virtual-list'
