+++ src/views/deck/mode-toolbar/page-settings.vue
@@ -3,7 +3,7 @@ import UiButton from '@/components/ui-kit/button.vue'
 import UiPopover from '@/components/ui-kit/popover.vue'
 import SectionList from '@/components/layout-kit/section-list.vue'
 import LabeledSection from '@/components/layout-kit/labeled-section.vue'
-import { deckViewShellKey, type CardGridSize } from '@/composables/deck/view-shell'
+import { deckViewShellKey, type CardGridSize } from '@/views/deck/composables/view-shell'
 import { inject, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { emitSfx } from '@/sfx/bus'
